pyproject.toml
[project]
name = "dataraum"
version = "0.1.0"
description = "Rich metadata context engine for AI-driven data analytics"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "Apache-2.0" }
authors = [
    { name = "DataRaum", email = "hello@dataraum.com" }
]
keywords = ["data", "metadata", "context", "ai", "analytics", "duckdb"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.14",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Data processing
    "duckdb>=1.0.0",
    "pint>=0.23",
    "pandas>=3.0.0",
    # Database (SQLAlchemy + SQLite backend)
    "sqlalchemy>=2.0.0",
    # Pydantic
    "pydantic>=2.0.0",
    # MCP
    "mcp>=1.0.0",
    # Configuration
    "pyyaml>=6.0.0",
    # Scientific computing
    "scipy>=1.11.0",
    # Time Series Analysis
    "statsmodels>=0.14.0",
    "ruptures>=1.1.0",
    # Utilities
    "structlog>=24.1.0",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "networkx>=3.6",
    "pgmpy>=0.1.25",
    "torch>=2.0.0",
    "anthropic>=0.76.0",
    # TUI
    "textual>=0.50.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.98.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
dataraum = "dataraum.cli:app"
dataraum-mcp = "dataraum.mcp.server:main"

[project.urls]
Homepage = "https://github.com/dataraum/dataraum-context"
Documentation = "https://docs.dataraum.com/context"
Repository = "https://github.com/dataraum/dataraum-context"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.98.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
    "pandas-stubs>=2.3.2.250926",
    "types-pyyaml>=6.0.12.20250915",
    "scipy-stubs>=1.16.3.2",
    "pytest-testmon>=2.2.0",
    "zensical>=0.0.24",
]
e2e = [
    "dataraum-testdata @ git+https://github.com/dataraum/dataraum-testdata.git",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
dataraum-testdata = { path = "../dataraum-testdata", editable = true }
torch = [{ index = "pytorch-cpu" }]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dataraum"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --ignore=tests/e2e"
markers = [
    "integration: marks tests as integration tests (may be slower)",
    "slow: marks tests as slow (use real data, skip in CI)",
    "e2e: end-to-end tests using dataraum-testdata (excluded from testmon)",
]
filterwarnings = [
    # Numpy warnings during correlation calculations with constant columns (expected)
    "ignore:invalid value encountered in divide:RuntimeWarning",
    "ignore:divide by zero encountered in scalar divide:RuntimeWarning",
    # DuckDB GIL warning on Python 3.14 free-threaded (external library)
    "ignore:The global interpreter lock.*:RuntimeWarning",
    # pgmpy invalid escape sequences (external library)
    "ignore:.*is an invalid escape sequence.*:SyntaxWarning",
]

[tool.ruff]
target-version = "py314"
line-length = 100
extend-exclude = [
    "prototypes/",
    "*.tcss",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "D",   # pydocstyle
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "D100",  # missing docstring in public module
    "D104",  # missing docstring in public package
    "D105",  # missing docstring in magic method
    "D106",  # missing docstring in public nested class
    "D107",  # missing docstring in __init__
    "D203",  # one-blank-line-before-class (conflicts with D211)
    "D213",  # multi-line-summary-second-line (conflicts with D212)
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"src/**" = [
    "D101", "D102", "D103",  # no backfill: missing docstrings in classes/methods/functions
    "D202", "D209", "D212",  # formatting in existing docstrings
    "D410", "D411", "D412",  # section blank line formatting
    "D415", "D417",          # punctuation and incomplete args
]
"tests/**" = ["D"]                    # no docstring enforcement in tests
"scripts/**" = ["D"]                  # no docstring enforcement in scripts

[tool.ruff.lint.isort]
known-first-party = ["dataraum"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_ignores = false  # Allow unused ignores during migration
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

# Ignore attr-defined errors for re-exported legacy types
# These are false positives due to exec()-based re-export in core.models.__init__
[[tool.mypy.overrides]]
module = "dataraum.core.models"

[[tool.mypy.overrides]]
module = [
    "duckdb.*",
    "mcp.*",
    "anthropic.*",
    "sdv.*",
    "scipy.*",
    "statsmodels.*",
    "ruptures.*",
    "networkx.*",
    "pgmpy.*",
    "textual.*",
    "typer.*",
]
ignore_missing_imports = true

# Unfinished features that depend on not-yet-implemented context module
[[tool.mypy.overrides]]
module = [
    "dataraum.llm.features.queries",
    "dataraum.llm.features.summary",
]
ignore_errors = true

# Test files don't need strict type checking
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
ignore_missing_imports = true
